requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.3.2
python-multipart==0.0.6